添加角色删除验证

+++ src/main/java/com/hnly/provincial/service/role/impl/RoleServiceImpl.java
@@ -3,14 +3,21 @@ package com.hnly.provincial.service.role.impl;
 import com.alibaba.druid.util.StringUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.hnly.provincial.comm.ResultEnum;
 import com.hnly.provincial.comm.utils.Conversion;
 import com.hnly.provincial.comm.utils.TableDataUtils;
+import com.hnly.provincial.config.interceptor.exception.MyException;
 import com.hnly.provincial.dao.role.RoleMapper;
 import com.hnly.provincial.entity.role.Role;
 import com.hnly.provincial.entity.role.RoleVO;
+import com.hnly.provincial.entity.role.menu.RoleMenu;
+import com.hnly.provincial.entity.user.User;
 import com.hnly.provincial.service.role.IRoleService;
+import com.hnly.provincial.service.role.menu.IRoleMenuService;
+import com.hnly.provincial.service.user.IUserService;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
 
@@ -24,6 +31,12 @@ import java.util.List;
  */
 @Service
 public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IRoleService {
+
+    @Resource
+    private IRoleMenuService roleMenuService;
+
+    @Resource
+    private IUserService userService;
 
     @Override
     public TableDataUtils<List<RoleVO>> findListByPage(RoleVO roleVO) {
@@ -44,7 +57,11 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
 
     @Override
     public boolean delete(Long id) {
+        if (userService.lambdaQuery().eq(User::getQuanxian, id).count() > 0) {
+            throw new MyException(ResultEnum.HIVE_NOT_DELETE);
+        }
         baseMapper.deleteById(id);
+        roleMenuService.lambdaUpdate().eq(RoleMenu::getRoleId, id).remove();
         return true;
     }
